🔨 Refactor Export (#2057)
* change Dockerfile to fix #1775

* Move export functionality to AnomalyModule

* Updated CHANGELOG

* replace flaky with pytest-rerunfailures


* update notebook


* resolved comments

* minor fixes

* revert to openvino-export via onnx

* fix changelog


* minor refactor


* move InferenceModel to a different file


* move export mixin to components


---------
