+++ tests/integration/model/test_models.py
@@ -12,7 +12,7 @@ import pytest
 
 from anomalib import TaskType
 from anomalib.data import AnomalibDataModule, MVTec
-from anomalib.deploy.export import ExportType
+from anomalib.deploy import ExportType
 from anomalib.engine import Engine
 from anomalib.models import AnomalyModule, get_available_models, get_model
